+++ src/TopicsViewer.cpp
@@ -6,16 +6,16 @@
 //======================================================================*/
 #include "TopicsViewer.h"
 
-TopicsViewer::TopicsViewer(Ui::DaqMain* ui, SignalsSlot* _topicSubscribers, QString _title, int x, int y)
-:DAQViz(ui, _title,x, y),
+TopicsViewer::TopicsViewer(Ui::DaqMain* ui, SignalsSlot* _topicSubscribers, QString _title, int x, int y,int frame)
+:DAQViz(ui, _title,x, y, frame),
 topicSubscribers(_topicSubscribers )
 {
+    interval_mode=false;
     DAQViz::buttonState=ButtonState::OFF;
     QObject::connect(topicSubscribers, &SignalsSlot::dataUpdated, this, &TopicsViewer::on_data_update_triggered);
     qlistView.setModel(&qstringList);
     //DAQViz::display_layout.addWidget(&qlistView);
     qlistView.hide();
-    viewInterval=0;
     ROS_INFO("DAQ: Topic viewer is ready: %s",qPrintable(_title) );
 }
 
@@ -24,7 +24,7 @@ TopicsViewer::~TopicsViewer() {
 }
 
 void TopicsViewer::showWindow() {
-    setViewInterval(0);
+    interval_mode=false;
     DAQViz::buttonState=ButtonState::ON;
     ui->rviz_layout->addWidget(DAQViz::title,0,0);
     ui->rviz_layout->addWidget(&qlistView,1,0);
@@ -35,8 +35,8 @@ void TopicsViewer::showWindow() {
 }
 
 
-void TopicsViewer::showWindow(int xpos, int ypos) {
-    setViewInterval(30);
+void TopicsViewer::showWindow(int xpos, int ypos, bool interval) {
+   interval_mode=interval;
     DAQViz::buttonState=ButtonState::ON;
     ui->rviz_layout->addWidget(DAQViz::title,xpos-1,ypos);
     ui->rviz_layout->addWidget(&qlistView,xpos,ypos);
@@ -45,9 +45,6 @@ void TopicsViewer::showWindow(int xpos, int ypos) {
     ui->rviz_layout->update();
 }
 
-void TopicsViewer::setViewInterval(int interval) {
-    viewInterval=interval;
-}
 
 void TopicsViewer::closeWindow() {
     DAQViz::buttonState=ButtonState::OFF;
@@ -64,16 +61,22 @@ void TopicsViewer::on_data_update_triggered(const QString& data) {
     //ROS_INFO("DAQ: Topic viewer is data: %s",qPrintable(data) );
     static int cnt=0;
     if (DAQViz::buttonState) {
-        if ( viewInterval == 0) {
-            QStringList list(data);
-            qstringList.setStringList(list);
-        }
-        else {
-            if ( cnt++ >= viewInterval ) {
+        if ( interval_mode) {
+            if ( cnt >= frame_cnt ) {
+                ROS_DEBUG("TopicsViewer::showWindow normal %d", cnt);
                 QStringList list(data);
                 qstringList.setStringList(list);
                 cnt=0;
             }
+            else {
+                ROS_DEBUG("TopicsViewer::showWindow normal--- %d", cnt);
+                cnt++;
+            }
+        }
+        else {
+            ROS_DEBUG("TopicsViewer::showWindow normal" );
+            QStringList list(data);
+            qstringList.setStringList(list);
         }
     }
 }